Change package from io.cloudsoft.terraform.template to io.cloudsoft.terraform.infrastructure
