+++ cloudsoft-terraform-template/cloudsoft-terraform-infrastructure.json
