+++ cloudsoft-terraform-template/pom.xml
@@ -46,6 +46,11 @@
             <groupId>software.amazon.awssdk</groupId>
             <artifactId>ssm</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.hierynomus</groupId>
+            <artifactId>sshj</artifactId>
+            <version>0.27.0</version>
+        </dependency>
 
         <!-- https://github.com/awslabs/aws-cloudformation-rpdk-java-plugin/ -->
         <dependency>
@@ -88,16 +93,6 @@
             <artifactId>mockito-junit-jupiter</artifactId>
             <version>2.26.0</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk-logs</artifactId>
-            <version>1.11.592</version>
-        </dependency>
-        <dependency>
-            <groupId>com.hierynomus</groupId>
-            <artifactId>sshj</artifactId>
-            <version>0.27.0</version>
         </dependency>
     </dependencies>
 
@@ -236,7 +231,7 @@
             <resource>
                 <directory>${project.basedir}</directory>
                 <includes>
-                    <include>cloudsoft-terraform-template.json</include>
+                    <include>cloudsoft-terraform-infrastructure.json</include>
                 </includes>
             </resource>
         </resources>
