+++ cloudsoft-terraform-template/resource-role.yaml
@@ -23,9 +23,9 @@ Resources:
             Statement:
               - Effect: Allow
                 Action:
-                - "ssm:GetParameter"
                 - "s3:GetObject"
                 - "s3:GetObjectVersion"
+                - "ssm:GetParameter"
                 Resource: "*"
 Outputs:
   ExecutionRoleArn:
