+++ cloudsoft-terraform-template/src/main/java/io/cloudsoft/terraform/infrastructure/CallbackContext.java
@@ -1,4 +1,4 @@
-package io.cloudsoft.terraform.template;
+package io.cloudsoft.terraform.infrastructure;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,9 +9,9 @@ public class CallbackContext {
 
     public String stepId;
     public int lastDelaySeconds;
-    
+
     // for testing:
     public boolean forceSynchronous = false;
     public boolean disregardCallbackDelay = false; // Call the worker again ASAP (use with care)
-    
+
 }
