+++ cloudsoft-terraform-template/src/main/java/io/cloudsoft/terraform/infrastructure/CreateHandler.java
@@ -1,6 +1,6 @@
-package io.cloudsoft.terraform.template;
+package io.cloudsoft.terraform.infrastructure;
 
-import io.cloudsoft.terraform.template.worker.DeleteHandlerWorker;
+import io.cloudsoft.terraform.infrastructure.worker.CreateHandlerWorker;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.ssm.SsmClient;
 import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
@@ -8,13 +8,14 @@ import software.amazon.cloudformation.proxy.Logger;
 import software.amazon.cloudformation.proxy.ProgressEvent;
 import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
-public class DeleteHandler extends TerraformBaseHandler<CallbackContext> {
+public class CreateHandler extends TerraformBaseHandler<CallbackContext> {
 
-    public DeleteHandler(SsmClient ssmClient, S3Client s3Client) {
+    // for tests
+    public CreateHandler(SsmClient ssmClient, S3Client s3Client) {
         super(ssmClient, s3Client);
     }
 
-    public DeleteHandler() {
+    public CreateHandler() {
         super();
     }
 
@@ -25,6 +26,6 @@ public class DeleteHandler extends TerraformBaseHandler<CallbackContext> {
             final CallbackContext callbackContext,
             final Logger logger) {
 
-        return run(callbackContext, cb -> new DeleteHandlerWorker(proxy, request, cb, logger, this));
+        return run(callbackContext, cb -> new CreateHandlerWorker(proxy, request, cb, logger, this));
     }
 }
