+++ cloudsoft-terraform-template/src/main/java/io/cloudsoft/terraform/infrastructure/DeleteHandler.java
@@ -1,6 +1,6 @@
-package io.cloudsoft.terraform.template;
+package io.cloudsoft.terraform.infrastructure;
 
-import io.cloudsoft.terraform.template.worker.CreateHandlerWorker;
+import io.cloudsoft.terraform.infrastructure.worker.DeleteHandlerWorker;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.ssm.SsmClient;
 import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
@@ -8,14 +8,13 @@ import software.amazon.cloudformation.proxy.Logger;
 import software.amazon.cloudformation.proxy.ProgressEvent;
 import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
 
-public class CreateHandler extends TerraformBaseHandler<CallbackContext> {
+public class DeleteHandler extends TerraformBaseHandler<CallbackContext> {
 
-    // for tests
-    public CreateHandler(SsmClient ssmClient, S3Client s3Client) {
+    public DeleteHandler(SsmClient ssmClient, S3Client s3Client) {
         super(ssmClient, s3Client);
     }
 
-    public CreateHandler() {
+    public DeleteHandler() {
         super();
     }
 
@@ -26,6 +25,6 @@ public class CreateHandler extends TerraformBaseHandler<CallbackContext> {
             final CallbackContext callbackContext,
             final Logger logger) {
 
-        return run(callbackContext, cb -> new CreateHandlerWorker(proxy, request, cb, logger, this));
+        return run(callbackContext, cb -> new DeleteHandlerWorker(proxy, request, cb, logger, this));
     }
 }
