+++ cloudsoft-terraform-template/src/test/java/io/cloudsoft/terraform/infrastructure/UpdateHandlerTest.java
@@ -1,4 +1,4 @@
-package io.cloudsoft.terraform.template;
+package io.cloudsoft.terraform.infrastructure;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
