+++ cloudsoft-terraform-template/src/test/java/io/cloudsoft/terraform/infrastructure/worker/AbstractHandlerWorkerTest.java
@@ -1,8 +1,8 @@
-package io.cloudsoft.terraform.template.worker;
+package io.cloudsoft.terraform.infrastructure.worker;
 
-import io.cloudsoft.terraform.template.CallbackContext;
-import io.cloudsoft.terraform.template.ResourceModel;
-import io.cloudsoft.terraform.template.TerraformBaseHandler;
+import io.cloudsoft.terraform.infrastructure.CallbackContext;
+import io.cloudsoft.terraform.infrastructure.ResourceModel;
+import io.cloudsoft.terraform.infrastructure.TerraformBaseHandler;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
