+++ cloudsoft-terraform-template/template.yml
@@ -10,13 +10,13 @@ Resources:
   TypeFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: io.cloudsoft.terraform.template.HandlerWrapper::handleRequest
+      Handler: io.cloudsoft.terraform.infrastructure.HandlerWrapper::handleRequest
       Runtime: java8
-      CodeUri: ./target/cloudsoft-terraform-template-handler-1.0-SNAPSHOT.jar
+      CodeUri: ./target/cloudsoft-terraform-infrastructure-handler-1.0-SNAPSHOT.jar
 
   TestEntrypoint:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: io.cloudsoft.terraform.template.HandlerWrapper::testEntrypoint
+      Handler: io.cloudsoft.terraform.infrastructure.HandlerWrapper::testEntrypoint
       Runtime: java8
-      CodeUri: ./target/cloudsoft-terraform-template-handler-1.0-SNAPSHOT.jar
+      CodeUri: ./target/cloudsoft-terraform-infrastructure-handler-1.0-SNAPSHOT.jar
